+++ lib/rollgeist/railtie.rb
@@ -5,9 +5,5 @@ module Rollgeist
     initializer "rollgeist.executor" do |application|
       ExecutionState.install!(application.executor)
     end
-
-    config.to_prepare do
-      Rollgeist.install_global_id_patch!
-    end
   end
 end